Changed Jms.session to only take acknowledgeMode, since it contains all the information needed.
You can only have transacted as true, if the acknowledge mode is 0,

and you can only use other acknowledge modes as long as transacted is false.

Therefore it is redundant to offer both parameters.

I've set a default value of AUTO_ACKNOWLEDGE, to keep consistent with previous versions and the spec in general.